+++ mod-pubsub-server/src/test/java/org/folio/rest/impl/PubSubAPITest.java
@@ -1,25 +1,202 @@
 package org.folio.rest.impl;
 
 import io.restassured.RestAssured;
+import io.restassured.response.Response;
+import io.vertx.core.json.JsonObject;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.apache.http.HttpStatus;
+import org.folio.rest.jaxrs.model.EventDescriptor;
+import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import static org.hamcrest.Matchers.empty;
+import static org.hamcrest.Matchers.is;
+
 @RunWith(VertxUnitRunner.class)
 public class PubSubAPITest extends AbstractRestTest {
+  
+  private static final String EVENT_TYPES_PATH = "/pubsub/event-types";
 
-  public static final String PUB_SUB_PATH = "/pubsub";
+  private EventDescriptor eventDescriptor = new EventDescriptor()
+    .withEventType("CREATED_SRS_MARC_BIB_RECORD_WITH_ORDER_DATA")
+    .withDescription("Created SRS Marc Bibliographic Record with order data in 9xx fields")
+    .withEventTTL(1)
+    .withSigned(false);
 
   @Test
-  public void shouldReturnOkOnGet() {
+  public void shouldReturnEmptyListOnGet() {
     RestAssured.given()
       .spec(spec)
       .when()
-      .get(PUB_SUB_PATH)
+      .get(EVENT_TYPES_PATH)
       .then()
-      .statusCode(HttpStatus.SC_OK);
+      .statusCode(HttpStatus.SC_OK)
+      .body("totalRecords", is(0))
+      .body("eventDescriptors", empty());
   }
 
+  @Test
+  public void shouldReturnAllEventDescriptorsOnGet() {
+    postEventDescriptor(eventDescriptor);
+
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .get(EVENT_TYPES_PATH)
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .body("totalRecords", is(1));
+  }
+
+  @Test
+  public void shouldReturnNotFoundOnGetByEventType() {
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .get(EVENT_TYPES_PATH + "/" + eventDescriptor.getEventType())
+      .then()
+      .statusCode(HttpStatus.SC_NOT_FOUND);
+  }
+
+  @Test
+  public void shouldReturnEventDescriptorOnGetByEventType() {
+    EventDescriptor createdEventDescriptor = postEventDescriptor(this.eventDescriptor);
+
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .get(EVENT_TYPES_PATH + "/" + createdEventDescriptor.getEventType())
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .body("eventType", is(this.eventDescriptor.getEventType()))
+      .body("description", is(this.eventDescriptor.getDescription()))
+      .body("eventTTL", is(this.eventDescriptor.getEventTTL()))
+      .body("signed", is(this.eventDescriptor.getSigned()));
+  }
+
+  @Test
+  public void shouldCreateEventDescriptorOnPost() {
+    RestAssured.given()
+      .spec(spec)
+      .body(eventDescriptor)
+      .when()
+      .post(EVENT_TYPES_PATH)
+      .then()
+      .statusCode(HttpStatus.SC_CREATED);
+  }
+
+  @Test
+  public void shouldReturnBadRequestOnPost() {
+    RestAssured.given()
+      .spec(spec)
+      .body(new JsonObject().toString())
+      .when()
+      .post(EVENT_TYPES_PATH)
+      .then()
+      .statusCode(HttpStatus.SC_UNPROCESSABLE_ENTITY);
+  }
+
+  @Test
+  public void shouldReturnBadRequestOnPostWithExistingEventType() {
+    EventDescriptor createdEventDescriptor = postEventDescriptor(this.eventDescriptor);
+
+    RestAssured.given()
+      .spec(spec)
+      .body(new EventDescriptor()
+        .withEventType(createdEventDescriptor.getEventType())
+        .withDescription("Test description")
+        .withEventTTL(1)
+        .withSigned(false))
+      .when()
+      .post(EVENT_TYPES_PATH)
+      .then()
+      .statusCode(HttpStatus.SC_UNPROCESSABLE_ENTITY);
+  }
+
+  @Test
+  public void shouldUpdateEventDescriptorOnPut() {
+    EventDescriptor eventDescriptor = new EventDescriptor()
+      .withEventType("CREATED_SRS_MARC_BIB_RECORD_WITH_ORDER_DATA")
+      .withDescription("Created SRS Marc Bibliographic Record with order data in 9xx fields")
+      .withEventTTL(1)
+      .withSigned(false);
+
+    EventDescriptor createdEventDescriptor = postEventDescriptor(eventDescriptor);
+    createdEventDescriptor.setDescription("New description");
+
+    RestAssured.given()
+      .spec(spec)
+      .body(createdEventDescriptor)
+      .when()
+      .put(EVENT_TYPES_PATH + "/" + createdEventDescriptor.getEventType())
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .body("eventType", is(createdEventDescriptor.getEventType()))
+      .body("description", is(createdEventDescriptor.getDescription()))
+      .body("eventTTL", is(createdEventDescriptor.getEventTTL()))
+      .body("signed", is(createdEventDescriptor.getSigned()));
+  }
+
+  @Test
+  public void shouldReturnBadRequestOnPut() {
+    RestAssured.given()
+      .spec(spec)
+      .body(new JsonObject().toString())
+      .when()
+      .put(EVENT_TYPES_PATH + "/" + eventDescriptor.getEventType())
+      .then()
+      .statusCode(HttpStatus.SC_UNPROCESSABLE_ENTITY);
+  }
+
+  @Test
+  public void shouldReturnNotFoundOnPut() {
+    RestAssured.given()
+      .spec(spec)
+      .body(eventDescriptor)
+      .when()
+      .put(EVENT_TYPES_PATH + "/" + eventDescriptor.getEventType())
+      .then()
+      .statusCode(HttpStatus.SC_NOT_FOUND);
+  }
+
+  @Test
+  public void shouldDeleteEventDescriptorOnDelete() {
+    EventDescriptor createdEventDescriptor = postEventDescriptor(eventDescriptor);
+
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .delete(EVENT_TYPES_PATH + "/" + createdEventDescriptor.getEventType())
+      .then()
+      .statusCode(HttpStatus.SC_NO_CONTENT);
+
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .get(EVENT_TYPES_PATH + "/" + createdEventDescriptor.getEventType())
+      .then()
+      .statusCode(HttpStatus.SC_NOT_FOUND);
+  }
+
+  @Test
+  public void shouldReturnNotFoundOnDelete() {
+    RestAssured.given()
+      .spec(spec)
+      .when()
+      .delete(EVENT_TYPES_PATH + "/" + "NONEXISTENT_EVENT_TYPE")
+      .then()
+      .statusCode(HttpStatus.SC_NOT_FOUND);
+  }
+
+  private EventDescriptor postEventDescriptor(EventDescriptor eventDescriptor) {
+    Response postResponse = RestAssured.given()
+      .spec(spec)
+      .body(eventDescriptor)
+      .when()
+      .post(EVENT_TYPES_PATH);
+    Assert.assertThat(postResponse.statusCode(), is(HttpStatus.SC_CREATED));
+    return postResponse.body().as(EventDescriptor.class);
+  }
 
 }